+++ src/lipsumpy/lipsum_cpp/__init__.py
@@ -1,3 +1,17 @@
+"""@package lipsum_cpp
+
+@brief Main file of lipsum-cpp Python wrapper
+
+This is the main file of lipsum-cpp's Python wrapper.
+This file is under the BSD Zero-Clause License.
+
+@example PyWrapper.py
+
+@copyright Copyright (c) 2026 LambBread
+
+@author LambBread from github.com
+"""
+
 # from ._ctypes_loader import native_lib
 import ctypes
 from ._low_level import native_lib
@@ -19,41 +33,164 @@ NOPORT = False
 PORT = True
 
 class CaseSlugCase:
+    """@brief Cases used by lipsum_cpp.Generator.case_slug()
+     
+     @since 0.5.4
+     
+     Types of cases used by lipsum_cpp.Generator.case_slug(). Includes various
+     types such as camel case, snake case, etc.
+    """
     CamelCase, PascalCase, SnakeCase, ShoutyCase, KebabCase, TrainCase = range(6)
 
 class CodeLanguage:
+    """@brief Languages used by lipsum_cpp.Generator.code()
+    
+     @since 0.5.4
+    
+     Programming languages that lipsum_cpp.Generator.code() can generate, such
+     as C++ and Python.
+    """
     Cpp, Python, Rust, C, JavaScript = range(5)
 
 class Format:
+    """@brief Types of formats.
+      
+     @since 0.5.4
+      
+     Types of formats used in lipsum_cpp.count_paragraphs() and
+     lipsum_cpp.convert_format(). In lipsum_cpp.count_paragraphs(), Plain is counting
+     plain text, Markdown is counting Markdown or HTML, and HTML is counting
+     HTML explicitly.
+    """
     Plain, Markdown, HTML, JSON, XML = range(5)
 
 class LipsumString:
+    """@brief Wrapper around new/delete C strings
+
+    @since 0.5.4
+
+    A string managed from the C wrapper using lpsm_DeleteString() and new/delete. It automatically
+    deletes strings and has conversion to Python strings.
+    """
     def __init__(self, ptr:int):
+        """@brief Constructor for LipsumString
+
+        @since 0.5.4
+
+        @param self The object pointer.
+        @param ptr The raw pointer handle to the string.
+        """
         self.__ptr = ptr
 
     def __str__(self):
+        """@brief Convert the LipsumString to a Python str
+
+        @since 0.5.4
+
+        Decode the LipsumString under UTF-8 as a Python str object.
+
+        @param self The object pointer.
+
+        @return str The Python string
+        """
         str_at = ctypes.string_at(self.__ptr)
         return str_at.decode("utf-8")
 
     def __del__(self):
+        """@brief Delete the LipsumString.
+
+        @since 0.5.4
+
+        Delete the LipsumString's handle using lpsm_DeleteString.
+
+        @param self The object pointer.
+        """
         native_lib.lpsm_DeleteString(self.__ptr)
 
 def generate_default_lipsum_sentence():
+    """@brief Generate the beginning Lorem Ipsum sentence.
+    
+     @since 0.5.4
+    
+     Returns the sentence "Lorem ipsum dolor sit amet, consectetur adipiscing
+     elit."
+    
+     @return LipsumString The default Lorem Ipsum sentence.
+    """
     return LipsumString(native_lib.lpsm_GenerateDefaultLipsumSentence())
 
 def count_words(str_:str):
+    """@brief Count the number of words in a string.
+    
+     @since 0.5.4
+    
+     Count the number of words in a string using alphanumeric characters,
+     hyphens, plus-signs, and apostrophes, excluding words within parentheses.
+    
+     @param str_ The string inputted.
+    
+     @return int The number of words.
+    """
     return native_lib.lpsm_CountWords(bytes(str_, "utf-8"))
 
 def count_sentence_fragments(str_:str):
+    """@brief Count the number of sentence fragments in a string.
+    
+     @since 0.5.4
+    
+     Count the number of commas, semicolons, hyphens, and colons in a string,
+     excluding usage in Markdown. (i.e. hyphen at start of line, or in
+     parentheses)
+    
+     @param str_ The string inputted.
+    
+     @return int The number of sentence fragments.
+     
+    """
     return native_lib.lpsm_CountSentenceFragments(bytes(str_, "utf-8"))
 
 def count_sentences(str_:str):
+    """@brief Count the number of sentences in a string.
+    
+     @since 0.5.4
+    
+     Count the number of periods, exclamation marks, and question marks in a
+     string, ignoring usage in Markdown URLs. (i.e. in parentheses)
+    
+     @param str_ The string inputted.
+    
+     @return int The number of sentences.
+    """
     return native_lib.lpsm_CountSentences(bytes(str_, "utf-8"))
 
 def count_paragraphs(str_:str, method:int = Format.Plain):
+    """@brief Count the number of paragraphs or elements in a string.
+    
+     @since 0.5.4
+    
+     Count the number of tab characters in a non-formatted string, double
+     newlines in Markdown/HTML strings, and elements in HTML strings.
+    
+     @param str_ The string inputted.
+     @param method The format the string is in. By default non-formatted.
+    
+     @return int The number of paragraphs. 
+    """
     return native_lib.lpsm_CountParagraphs(bytes(str_, "utf-8"), method)
 
 def convert_format(str_:str, format1:int = Format.Plain, format2:int = Format.HTML):
+    """@brief Convert a string between two formats.
+    
+     @since 0.5.4
+    
+     Attempt a conversion of a string between the two specified formats.
+    
+     @param str_ The string inputted.
+     @param format1 The string's format. By default plain text.
+     @param format2 The destination format. By default HTML.
+    
+     @return str The converted string.
+    """
     return LipsumString(native_lib.lpsm_ConvertFormat(bytes(str_, "utf-8"), format1, format2))
 
 class Generator: