+++ src/lipsumpy/lipsum_cpp/custom_exceptions.py
@@ -1,5 +1,18 @@
-"""Exceptions defined for lipsum-cpp."""
+"""@file custom_exceptions.py
+
+@brief Exceptions defined for lipsum-cpp.
+
+This file defines exceptions for lipsum-cpp.
+This file is under the BSD Zero-Clause License.
+
+@copyright Copyright (c) 2026 LambBread
+
+@author LambBread from github.com
+"""
 
 class LipsumCppException(Exception):
-    """An exception originating from the C++ library."""
+    """@brief An exception originating from the C++ library.
+    
+    @since 0.5.4
+    """
     pass